Changes the ways roles work in the database
This creates a roles and a userrole table.
The roles table will hold the different roles that users can have, i.e. Student, Faculty, Admin.
The userrole table will hold the userid and the roleid.
This will allow us to put multiple roles on one user incase someone is an admin and a faculty member.
There are also some route edits to make the different routing work to actually use the tables.