Adds accompanist and instructor link table
Adds an accompanist and instructor link table to the database. This table will be used to link accompanists and instructors to a particular student so that a student can have multiple accompanists and instructors.

- Added an accompanist table that only has an ID for now
- Added a studentinstructor table with an id and two foreign keys for the studentinfo table and the instructors table
- Added a studentaccompanist table with an id and two foreign keys for the studentinfo table and the accompanist table
- Added a controller for the studentinstructor table that has a findStudent method and a findInstructor method
- Added a controller for the studentaccompanist table that has a findStudent method and a findAccompanist method

We can add more methods as needed.
We can also change the accompanist model as we start adding an accompanist role.