Add Functionality to Send Emails
This adds functionality to send emails from the backend.
There are currently two function to send emails and they are inside of the composer.js file

Send Notification
- This send a notification that a critique has been done for a student on a an event

Send Message
- This allows someone to send a custom message and subject to a custom list of students that are all bcc'ed for privacy

This is kind of piece mealed together from different documentation that I found because no one has documentation for the exact way that I was trying to do it because I was trying to keep it secure and make it work without saving a file to the backend and letting it work for more than one person.

In order to get this to work locally, you will need to add the gmail api client credentials to your .env file, but they have already been added as secrets to GitHub to work on Prod